backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
supabase>=2.9.0
langchain>=0.3.7
langchain-groq>=0.2.1
langchain-core>=0.3.19
langgraph>=0.2.44
e2b-code-interpreter>=1.0.3
websockets>=13.1
httpx>=0.27.0
python-frontmatter>=1.1.0
